server/requirements.txt
paho-mqtt
pymysql
omegaconf